Fixed some bugs, added redis db

+++ src/utils/sqs.js
@@ -1,5 +1,7 @@
 const AWS = require('aws-sdk');
 const getAllLinksInPage = require('./cheerio');
+const Site = require('./site');
+const { setNewSiteInRedis, getSiteLinksFromRedis } = require('../utils/redis');
 
 const sqs = new AWS.SQS({
     apiVersion: '2012-11-05',
@@ -8,6 +10,10 @@ const sqs = new AWS.SQS({
 
 const sendMessageToQueue = async (QueueUrl, url, level) => {
     console.log(url, level);
+    if (level === 0) {
+        messagesCounter = 0;
+        urlArr = [];
+    }
     try {
         const { MessageId } = await sqs.sendMessage({
             QueueUrl,
@@ -25,7 +31,7 @@ const sendMessageToQueue = async (QueueUrl, url, level) => {
         return MessageId;
     } catch (err) {
         console.log(err);
-        throw new Error(err);
+        throw new Error(err.message);
     }
 }
 
@@ -68,16 +74,28 @@ const deleteMessagesFromQueue = async (QueueUrl, messages) => {
     }
 }
 
+const deleteQueue = async (QueueUrl) => {
+    try {
+        await sqs.deleteQueue({ QueueUrl }).promise();
+        console.log('deleted');
+    } catch (err) {
+        throw new Error(err.message);
+    }
+}
 
-let receiveMessagesProcessResolved = { isTrue: true };
 
-const handleMessagesFromQueue = async (queueUrl, maxDepth, searchInterval) => {
+let receiveMessagesProcessResolved = { isTrue: true }; //Has to be object (instead of simple bool value)
+let messagesCounter;
+let urlArr = [];
+
+const handleMessagesFromQueue = async (queueUrl, maxDepth, maxPages, clearSearchInterval) => {
     receiveMessagesProcessResolved.isTrue = false;
     try {
         const messages = await pollMessagesFromQueue(queueUrl);
 
+        // Is search complete
         if (messages.length === 0) {
-            clearInterval(searchInterval);
+            await clearSearchInterval();
             console.log('Search Complete');
             return;
         }
@@ -85,18 +103,57 @@ const handleMessagesFromQueue = async (queueUrl, maxDepth, searchInterval) => {
         console.log('new poll batch');
 
         for (let message of messages) {
+            if (receiveMessagesProcessResolved.isTrue) break; // Queue deleted before processing finished
+
             console.log();
             console.log('message');
-            let currentMessageLevel = parseInt(message.MessageAttributes.level.StringValue);
-            if (currentMessageLevel >= maxDepth) break;
 
-            let links = await getAllLinksInPage(message.Body);
+            let currentMessageLevel = parseInt(message.MessageAttributes.level.StringValue);
+            if (currentMessageLevel >= maxDepth) {
+                console.log('Reached max depth');
+                break;
+            }
+            if (messagesCounter >= maxPages) {
+                console.log('Reached max pages');
+                break;
+            }
+
+            if (urlArr.includes(message.Body)) break;
+            urlArr.push(message.Body);
+
+            let shouldSaveSiteOnRedis = false;
+
+            let links = await getSiteLinksFromRedis(message.Body); //
+            if (links == undefined || links.length === 0) {
+                shouldSaveSiteOnRedis = true;
+                links = await getAllLinksInPage(message.Body);
+            }
 
             for (let link of links) {
+                if (receiveMessagesProcessResolved.isTrue) {
+                    shouldSaveSiteOnRedis = false;
+                    break; // Queue deleted before processing finished
+                }
+                
                 await sendMessageToQueue(queueUrl, link, currentMessageLevel + 1);
+                
+                new Site(link, currentMessageLevel + 1, message.Body);
+
+                messagesCounter++;
+
+                if (messagesCounter >= maxPages) {
+                    shouldSaveSiteOnRedis = false;
+                    break;
+                }
+            }
+
+            if (shouldSaveSiteOnRedis) {
+                await setNewSiteInRedis(message.Body, links); //
             }
         }
 
+        if (receiveMessagesProcessResolved.isTrue) return; // Queue deleted before processing finished
+        
         await deleteMessagesFromQueue(queueUrl, messages);
 
         receiveMessagesProcessResolved.isTrue = true;
@@ -111,6 +168,7 @@ const handleMessagesFromQueue = async (queueUrl, maxDepth, searchInterval) => {
 module.exports = {
     sqs,
     sendMessageToQueue,
+    deleteQueue,
     handleMessagesFromQueue,
     receiveMessagesProcessResolved
 };
